incident-agent/requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Async support
asyncio>=3.4.3
aiohttp>=3.9.0

# File watching
watchdog>=4.0.0

# Slack
slack-sdk>=3.27.0
slack-bolt>=1.18.0

# Database
sqlite3>=0.0.0
chromadb>=0.4.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0
